Refactor step interface and update documentation in corrector module
- Replaced `_StepInterface`, `_PlainStepInterface`, and `_ArmijoStepInterface` with `_PlainStep` and `_ArmijoStep` to simplify the step interface structure.
- Updated documentation to reflect the new class names and their functionalities, ensuring consistency across the corrector module.
- Removed obsolete `_step_interface.py` and consolidated step-related functionality into the new `stepping` module for better organization and maintainability.
- Adjusted import paths in relevant files to align with the new structure, enhancing code clarity and usability.
